% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_dating.R
\name{add_all_periods_to_all_contexts}
\alias{add_all_periods_to_all_contexts}
\title{add_all_periods_to_all_contexts}
\usage{
add_all_periods_to_all_contexts(data)
}
\arguments{
\item{data}{}
}
\description{
add_all_periods_to_all_contexts
}
