% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_plots.R
\name{add_caption}
\alias{add_caption}
\title{add caption to plotly layout in upper right corner}
\usage{
add_caption(plotly_fig, caption = "")
}
\arguments{
\item{caption}{chr (content of caption)}
}
\value{
a plotly plot
}
\description{
add caption to plotly layout in upper right corner
}
