% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_plots.R
\name{add_legend}
\alias{add_legend}
\title{use bottom legend}
\usage{
add_legend(plotly_fig, title = "", position = "bottom", mb = 100, y = -0.2)
}
\arguments{
\item{legend_title}{chr (content of title)}
}
\value{
a plotly plot
}
\description{
use bottom legend
}
