% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_plots.R
\name{add_title}
\alias{add_title}
\title{add title to plotly layout in upper left corner}
\usage{
add_title(plotly_fig, title = "")
}
\arguments{
\item{title}{chr (content of title)}
}
\value{
a plotly plot
}
\description{
add title to plotly layout in upper left corner
}
