% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_ui_elements.R
\name{addLRPopover}
\alias{addLRPopover}
\title{Popover for 'Load Resources'-buttons}
\usage{
addLRPopover(input)
}
\arguments{
\item{input}{the button}
}
\value{
same input with a tooltip/popover to inform of 'Load Resources'-button behaviour
}
\description{
Popover for 'Load Resources'-buttons
}
