% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_get_data.R
\name{alt_gather_trenches}
\alias{alt_gather_trenches}
\title{Alternative to gather trenches for milQuant}
\usage{
alt_gather_trenches(uidlist)
}
\arguments{
\item{uidlist}{index of the project db}
}
\value{
the index with Places
}
\description{
Alternative to gather trenches for milQuant
}
\examples{
connection <- connect_idaifield(project = "milet", pwd = "hallo")
uidlist <- get_field_index(connection)
}
