% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{UI for the milQuant app}
\usage{
app_ui(app_version = packageVersion("milQuant"))
}
\arguments{
\item{app_version}{Version of the Package to be display in UI}
}
\value{
the UI
}
\description{
UI for the milQuant app
}
