% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_global_helpers.R
\name{convert_to_Plotly}
\alias{convert_to_Plotly}
\title{convert_to_Plotly}
\usage{
convert_to_Plotly(
  ggplot_p,
  source = "ggplot_source",
  tooltip = c("y", "x", "fill")
)
}
\arguments{
\item{tooltip}{}
}
\description{
convert_to_Plotly
}
