% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_ui_elements.R
\name{create_login_modal}
\alias{create_login_modal}
\title{Produces the Modal that displays the Login/Connection Settings}
\usage{
create_login_modal()
}
\value{
a modal UI object
}
\description{
Produces the Modal that displays the Login/Connection Settings
}
