% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_activity.R
\name{db_activity_server}
\alias{db_activity_server}
\title{Tab-Module for Activity (Server Code)}
\usage{
db_activity_server(id)
}
\arguments{
\item{id}{The namespace-id for this Module.}
}
\value{
Server Code for the Module
}
\description{
Tab-Module for Activity (Server Code)
}
