% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_activity.R
\name{db_activity_tab}
\alias{db_activity_tab}
\title{Title}
\usage{
db_activity_tab(id, tabname)
}
\arguments{
\item{id}{Session id}

\item{tabname}{Name of the tab used to link it}
}
\value{
A tab
}
\description{
Title
}
