% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_home_db_overview.R
\name{db_overview_server}
\alias{db_overview_server}
\title{TabPanel-Module for Project Overview (Server Code)}
\usage{
db_overview_server(id)
}
\arguments{
\item{id}{The namespace-id for this Module.}
}
\value{
Server Code for the Module
}
\description{
TabPanel-Module for Project Overview (Server Code)
}
