% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_home_db_overview.R
\name{db_overview_tab}
\alias{db_overview_tab}
\title{Title}
\usage{
db_overview_tab(id, tabname)
}
\arguments{
\item{id}{Session id}

\item{tabname}{Name of the tab used to link it}
}
\value{
tabPanel with overview of the project db
}
\description{
Title
}
