% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_dating.R
\name{derive_dating_from_periods}
\alias{derive_dating_from_periods}
\title{derive_dating_from_periods}
\usage{
derive_dating_from_periods(data)
}
\arguments{
\item{data}{data.frame with the resources to be transformed}
}
\value{
data.frame with absolute values for dating where they were empty previously
}
\description{
#TODO use this from chrongler
}
