% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_dev.R
\name{devel_idf_version}
\alias{devel_idf_version}
\title{For checking while in development}
\usage{
devel_idf_version()
}
\value{
TRUE if there is no problem, FALSE if there is one
}
\description{
For checking while in development
}
\keyword{internal}
