% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_global_downloadModule.R
\name{downloadPlotButtons}
\alias{downloadPlotButtons}
\title{Global Module: downloadPlotButtons (UI)}
\usage{
downloadPlotButtons(id)
}
\arguments{
\item{id}{The namespace-id for this Module.}
}
\value{
A \code{div()} with a \code{downloadButton()}
}
\description{
Global Module: downloadPlotButtons (UI)
}
