% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_global_layerSelection.R
\name{generateLayerSelector}
\alias{generateLayerSelector}
\title{Title}
\usage{
generateLayerSelector(id, data, inputid)
}
\arguments{
\item{id}{Session id}

\item{data}{all the data (this is excessive...)}

\item{inputid}{the inputid the selector should have}
}
\value{
server code for layer selector
}
\description{
Title
}
