% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_global_periodSelection.R
\name{generatePeriodSelector}
\alias{generatePeriodSelector}
\title{Global Module: periodSelector (Server Code)}
\usage{
generatePeriodSelector(id, inputid)
}
\arguments{
\item{id}{The namespace-id for this Module.}

\item{inputid}{The \code{inputId} that should be used for the UI-Element.
Specified here to avoid \code{ns()}-problems and keep track of the value easier.}
}
\value{
Server Code for the Module
}
\description{
Global Module: periodSelector (Server Code)
}
