% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_global_get_data.R
\name{get_list_of_operations_in_places}
\alias{get_list_of_operations_in_places}
\title{Compiles a vector of Operations in the selected Place}
\usage{
get_list_of_operations_in_places(index, selected_places)
}
\arguments{
\item{selected_operations}{}
}
\value{
a vector of operation identifiers
}
\description{
Uses the index to build said vector. Filters for Operations that are
children of the selected places.
}
