% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_get_data.R
\name{get_list_of_places_with_finds}
\alias{get_list_of_places_with_finds}
\title{Compiles a vector of Places which contain finds}
\usage{
get_list_of_places_with_finds(index)
}
\arguments{
\item{index}{}
}
\value{
a vector of Place identifiers
}
\description{
Uses the index to build said vector. This way, Places which do not
contain any finds are not listed in the UI-element. milQuant obviously
needs find- and quantification-resources to work.
}
