% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_plots.R
\name{get_plot_vars}
\alias{get_plot_vars}
\title{Title}
\usage{
get_plot_vars(resource_category, colnames, type = "categorical")
}
\arguments{
\item{resource_category}{category of resources to get vars for}

\item{colnames}{colnames in the data}

\item{type}{which sort of vars for plotting should be selected}
}
\value{
a vector of possible variables
}
\description{
get the plottable variables from the config to preselect for fields that may otherwise be impossible to plot
}
