% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_get_data.R
\name{idf_uid_query}
\alias{idf_uid_query}
\title{idf_uid_query}
\usage{
idf_uid_query(login_connection, uids)
}
\arguments{
\item{login_connection}{connect_idaifield()-object}

\item{uids}{vector of UUIDs that should be downloaded}
}
\value{
a docs list with the resources of said UUIDs
}
\description{
idf_uid_query
}
