% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_modals.R
\name{make_busy_dialog}
\alias{make_busy_dialog}
\title{Produce the modal which is used to indicate that the Index is loaded}
\usage{
make_busy_dialog(project = "")
}
\arguments{
\item{project}{name of the project}
}
\value{
UI-Element modalDialog()
}
\description{
The modal pops up on Startup and prompts
IP-Adress and Password (as well as username,
but that is currently optional.)
}
