% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_tab_home.R
\name{make_home_tab}
\alias{make_home_tab}
\title{Make the UI for the landing page ("home")}
\usage{
make_home_tab(tabname = "tab_home")
}
\arguments{
\item{tabname}{Name of the tab used to link it}
}
\value{
a tabitem()
}
\description{
Make the UI for the landing page ("home")
}
