% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_modals.R
\name{make_login_dialog}
\alias{make_login_dialog}
\title{Produce the UI of the Login modal}
\usage{
make_login_dialog()
}
\value{
UI-Element modalDialog()
}
\description{
The modal pops up on Startup and prompts
IP-Adress and Password (as well as username,
but that is currently optional.)
}
