% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_global_downloadModule.R
\name{makeDownloadPlotHandler}
\alias{makeDownloadPlotHandler}
\title{Global Module: downloadPlotButtons (Server Code)}
\usage{
makeDownloadPlotHandler(id, dlPlot)
}
\arguments{
\item{id}{The namespace-id for this Module.}

\item{dlPlot}{The plot to be downloaded.}
}
\value{
Server Code for the Module
}
\description{
Global Module: downloadPlotButtons (Server Code)
}
