% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milQuant-package.R
\docType{package}
\name{milQuant-package}
\alias{milQuant}
\alias{milQuant-package}
\title{milQuant: Basic Quantitative Visualizations of Data from Field Desktop for Find Processing in Miletus}
\description{
The milQuant-Shiny Dashboard provides quick quantitative overviews of the data in a Field Desktop-project using the milet-configuration. It connects to the database automatically and can plot various graphs to provide quantitative visualizations focused on find-resources.
}
\author{
\strong{Maintainer}: Lisa Steinmann \email{lisa.steinmann@rub.de} (\href{https://orcid.org/0000-0002-2215-1243}{ORCID}) [copyright holder]

}
\keyword{internal}
