% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{milQuant_cats}
\alias{milQuant_cats}
\title{Resource categories for milQuant}
\format{
\subsection{\code{milQuant_cats}}{

A list:
\describe{
\item{one}{bla}
\item{all other values}{blabla}
...
}
}
}
\source{
Field Desktop
}
\usage{
milQuant_cats
}
\description{
Resource categories for milQuant
}
\keyword{datasets}
