% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_utility.R
\name{milQuant_hovertemplate}
\alias{milQuant_hovertemplate}
\title{Title}
\usage{
milQuant_hovertemplate(value = "count")
}
\arguments{
\item{value}{what should it say in the third row? (count / weight etc)}
}
\value{
character vector with hovertemplare for plotly
}
\description{
Title
}
