% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_plots.R
\name{milquant_plotly}
\alias{milquant_plotly}
\title{milquant_plotly}
\usage{
milquant_plotly(plotly_fig)
}
\arguments{
\item{plotly_fig}{a plotly plot}
}
\value{
a plotly plot
}
\description{
general layout parameters for all plotly plots
}
