% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_plots.R
\name{milquant_plotly_layout}
\alias{milquant_plotly_layout}
\title{milquant_plotly_layout}
\usage{
milquant_plotly_layout(plotly_fig, caption = FALSE)
}
\arguments{
\item{caption}{}
}
\description{
milquant_plotly_layout
}
