% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_finds_aoristic.R
\name{mod_aoristic_finds_serv}
\alias{mod_aoristic_finds_serv}
\title{Tab-Module for Aoristic Plots (Server Code)}
\usage{
mod_aoristic_finds_serv(id)
}
\arguments{
\item{id}{The namespace-id for this Module.}
}
\value{
Server Code for the Module
}
\description{
Tab-Module for Aoristic Plots (Server Code)
}
