% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_finds_aoristic.R
\name{mod_aoristic_finds_ui}
\alias{mod_aoristic_finds_ui}
\title{Tab-Module for Aoristic Plots (UI)}
\usage{
mod_aoristic_finds_ui(id, tabname)
}
\arguments{
\item{id}{The namespace-id for this Module.}

\item{tabname}{Name of the tab used to link this module.}
}
\value{
UI for the Module
}
\description{
Tab-Module for Aoristic Plots (UI)
}
