% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_finds_barchart.R
\name{mod_barchart_finds_serv}
\alias{mod_barchart_finds_serv}
\title{Tab-Module for \emph{Find}-Bar Charts (Server Code)}
\usage{
mod_barchart_finds_serv(id)
}
\arguments{
\item{id}{The namespace-id for this Module.}
}
\value{
Server Code for the Module
}
\description{
Tab-Module for \emph{Find}-Bar Charts (Server Code)
}
