% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_finds_barchart.R
\name{mod_barchart_finds_ui}
\alias{mod_barchart_finds_ui}
\title{Title}
\usage{
mod_barchart_finds_ui(id, tabname)
}
\arguments{
\item{id}{Session id}

\item{tabname}{Name of the tab used to link it}
}
\value{
The tab with generalized bar charts where categories can be selected
}
\description{
Title
}
