% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_finds_overview.R
\name{mod_finds_overview_serv}
\alias{mod_finds_overview_serv}
\title{Tab-Module for Overview of Inventoried Finds (Server Code)}
\usage{
mod_finds_overview_serv(id)
}
\arguments{
\item{id}{The namespace-id for this Module.}
}
\value{
Server Code for the Module
}
\description{
Tab-Module for Overview of Inventoried Finds (Server Code)
}
