% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_finds_overview.R
\name{mod_finds_overview_ui}
\alias{mod_finds_overview_ui}
\title{Tab-Module for Overview of Inventoried Finds (UI)}
\usage{
mod_finds_overview_ui(id, tabname)
}
\arguments{
\item{id}{The namespace-id for this Module.}

\item{tabname}{Name of the tab used to link this module.}
}
\value{
UI for the Module
}
\description{
Tab-Module for Overview of Inventoried Finds (UI)
}
