% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_finds_quant.R
\name{mod_finds_quant_serv}
\alias{mod_finds_quant_serv}
\title{Tab-Module for all Quantification Forms (Server Code)}
\usage{
mod_finds_quant_serv(id)
}
\arguments{
\item{id}{The namespace-id for this Module.}
}
\value{
Server Code for the Module
}
\description{
Tab-Module for all Quantification Forms (Server Code)
}
