% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_finds_quant.R
\name{mod_finds_quant_ui}
\alias{mod_finds_quant_ui}
\title{Tab-Module for all Quantification Forms (UI)}
\usage{
mod_finds_quant_ui(id, tabname)
}
\arguments{
\item{id}{The namespace-id for this Module.}

\item{tabname}{Name of the tab used to link this module.}
}
\value{
UI for the Module
}
\description{
#TODO: add up all individual counts / weights in case total does not exist;
but could also do that more reasonable by exporting and overwriting in the db itself.
}
