% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_loomweights_hist.R
\name{mod_loomweights_hist_serv}
\alias{mod_loomweights_hist_serv}
\title{Tab-Module for Loomweight-Historgram (Server Code)}
\usage{
mod_loomweights_hist_serv(id)
}
\arguments{
\item{id}{The namespace-id for this Module.}
}
\value{
Server Code for the Module
}
\description{
Tab-Module for Loomweight-Historgram (Server Code)
}
