% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_loomweights_hist.R
\name{mod_loomweights_hist_ui}
\alias{mod_loomweights_hist_ui}
\title{Title}
\usage{
mod_loomweights_hist_ui(id, tabname)
}
\arguments{
\item{id}{Session id}

\item{tabname}{Name of the tab used to link it}
}
\value{
A tab which displays histogram of weights for loomweights, curated
}
\description{
Title
}
