% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_pottery_QA.R
\name{mod_pottery_QA_serv}
\alias{mod_pottery_QA_serv}
\title{Tab-Module for Pottery Quantification A (Server Code)}
\usage{
mod_pottery_QA_serv(id)
}
\arguments{
\item{id}{The namespace-id for this Module.}
}
\value{
Server Code for the Module
}
\description{
Tab-Module for Pottery Quantification A (Server Code)
}
