% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_pottery_QB.R
\name{mod_pottery_QB_ui}
\alias{mod_pottery_QB_ui}
\title{Tab-Module for Pottery Quantification B (UI)}
\usage{
mod_pottery_QB_ui(id, tabname)
}
\arguments{
\item{id}{The namespace-id for this Module.}

\item{tabname}{Name of the tab used to link this module.}
}
\value{
UI for the Module
}
\description{
Tab-Module for Pottery Quantification B (UI)
}
\section{TODO might want to make it possible to do this with absolute}{
values replaced from the periods
}

