% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_pottery_single.R
\name{mod_pottery_single_ui}
\alias{mod_pottery_single_ui}
\title{Tab-Module for Single Pottery Resources (UI)}
\usage{
mod_pottery_single_ui(id, tabname)
}
\arguments{
\item{id}{The namespace-id for this Module.}

\item{tabname}{Name of the tab used to link this module.}
}
\value{
UI for the Module
}
\description{
#TODO make this less dynanic and more curated
}
