% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_quants.R
\name{mod_quants_serv}
\alias{mod_quants_serv}
\title{Title}
\usage{
mod_quants_serv(id, resource_category = "Brick_Quantification")
}
\arguments{
\item{id}{Session id}

\item{resource_category}{the category of resources to display in this tab}
}
\value{
server code
}
\description{
Title
}
