% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_quants.R
\name{mod_quants_ui}
\alias{mod_quants_ui}
\title{Title}
\usage{
mod_quants_ui(id, tabname)
}
\arguments{
\item{id}{Session id}

\item{tabname}{Name of the tab used to link it}
}
\value{
Generalized tab for Quantification forms, should work for all of them
}
\description{
Title
}
