% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_search.R
\name{mod_search_serv}
\alias{mod_search_serv}
\title{Tab-Module for Fulltext Search (Server Code)}
\usage{
mod_search_serv(id)
}
\arguments{
\item{id}{The namespace-id for this Module.}
}
\value{
Server Code for the Module
}
\description{
Tab-Module for Fulltext Search (Server Code)
}
