% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_search.R
\name{mod_search_ui}
\alias{mod_search_ui}
\title{Tab-Module for Fulltext Search (UI)}
\usage{
mod_search_ui(id, tabname = NULL)
}
\arguments{
\item{id}{The namespace-id for this Module.}

\item{tabname}{Name of the tab used to link this module.}
}
\value{
UI for the Module
}
\description{
Tab-Module for Fulltext Search (UI)
}
