% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milQuant-package.R
\name{mod_serv_doc}
\alias{mod_serv_doc}
\title{The Server-Code of a Module}
\usage{
mod_serv_doc(id, resource_category)
}
\arguments{
\item{id}{The namespace-id for this Module.}

\item{resource_category}{The category of \emph{resource}s to be used in this module.}

\item{inputid}{The \code{inputId} that should be used for the UI-Element.
Specified here to avoid \code{ns()}-problems and keep track of the value easier.}
}
\value{
Server Code for the Module
}
\description{
The Server-Code of a Module
}
\keyword{internal}
