% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milQuant-package.R
\name{mod_ui_doc}
\alias{mod_ui_doc}
\title{The UI of a Module}
\usage{
mod_ui_doc(id, tabname)
}
\arguments{
\item{id}{The namespace-id for this Module.}

\item{tabname}{Name of the tab used to link this module.}
}
\value{
UI for the Module
}
\description{
The UI of a Module
}
\keyword{internal}
