% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_workflow.R
\name{mod_worflow_serv}
\alias{mod_worflow_serv}
\title{Tab-Module for Workflow Overview (Server Code)}
\usage{
mod_worflow_serv(id)
}
\arguments{
\item{id}{The namespace-id for this Module.}
}
\value{
Server Code for the Module
}
\description{
Tab-Module for Workflow Overview (Server Code)
}
