% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_workflow.R
\name{mod_worflow_ui}
\alias{mod_worflow_ui}
\title{Title}
\usage{
mod_worflow_ui(id, tabname = NULL)
}
\arguments{
\item{id}{Session id}

\item{tabname}{Name of the tab used to link it}
}
\value{
Tab that display the workflow status selectable by layer and category
}
\description{
Title
}
