% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_ui_elements.R
\name{mq_spinner}
\alias{mq_spinner}
\title{Add a spinning milQuant-Logo}
\usage{
mq_spinner(object)
}
\arguments{
\item{object}{ui objects that needs a spinner}
}
\value{
ui object with spinner
}
\description{
Add a spinning milQuant-Logo
}
