% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_utility.R
\name{ms_days}
\alias{ms_days}
\title{Days to Milliseconds for Plotly Histogram bins}
\usage{
ms_days(days)
}
\arguments{
\item{days}{number of days}
}
\value{
\code{days} in milliseconds
}
\description{
Days to Milliseconds for Plotly Histogram bins
}
