% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_dating.R
\name{multiply_resources_by_period_range}
\alias{multiply_resources_by_period_range}
\title{multiply_resources_by_period_range}
\usage{
multiply_resources_by_period_range(data, split = TRUE)
}
\arguments{
\item{data}{}
}
\description{
multiply_resources_by_period_range
}
