% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_global_helpers.R
\name{period_filter}
\alias{period_filter}
\title{apply the period filter if the config is milet}
\usage{
period_filter(find_df, is_milet = FALSE, selector = NULL)
}
\arguments{
\item{selector}{}
}
\description{
apply the period filter if the config is milet
}
