% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_global_displayPlotDataTable.R
\name{plotDataTable_server}
\alias{plotDataTable_server}
\title{Title}
\usage{
plotDataTable_server(id, resources, click_data, x, customdata)
}
\arguments{
\item{id}{Session id}

\item{resources}{data to be display in the table}

\item{click_data}{the plotly click data}

\item{x}{first variable to select for}

\item{customdata}{second variable to select for}
}
\value{
server side for the datatable
}
\description{
Title
}
