% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_global_displayPlotDataTable.R
\name{plotDataTable_ui}
\alias{plotDataTable_ui}
\title{Title}
\usage{
plotDataTable_ui(id)
}
\arguments{
\item{id}{Session id}
}
\value{
a fluidRow with column selector, h2 and the datatable output
}
\description{
Title
}
