% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_ui_elements.R
\name{plotTitleInputs}
\alias{plotTitleInputs}
\title{tagList of two textInputs for plot title}
\usage{
plotTitleInputs(id = id, title = "", subtitle = "")
}
\arguments{
\item{title}{chr. default value for title textInput}

\item{subtitle}{chr. default value for subtitle textInput}
}
\value{
tagList with two textInput() UI elements
}
\description{
inputIDs are always \code{ns("title")} and \code{ns("subtitle")}
}
