% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_get_data.R
\name{prep_for_shiny}
\alias{prep_for_shiny}
\title{Prep for shiny}
\usage{
prep_for_shiny(data, reorder_periods = reorder_periods)
}
\arguments{
\item{data}{data (list) from get_idaifield_docs}

\item{reorder_periods}{should periods be an ordered factor? (only if is_milet)}
}
\value{
a data.frame
}
\description{
Prep for shiny
}
