% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_utility.R
\name{read_milQuant_settings}
\alias{read_milQuant_settings}
\title{Read the selection settings from the sys.file}
\usage{
read_milQuant_settings()
}
\value{
a list with the settings
}
\description{
This is used to make the selection settings restorable (i.e. which
Operation / Place / project was selected when milQuant was closed using
the save & close button)
}
