% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_utility.R
\name{remove_na_cols}
\alias{remove_na_cols}
\title{Helper to remove columns that are empty}
\usage{
remove_na_cols(data)
}
\arguments{
\item{data}{}
}
\description{
Helper to remove columns that are empty
}
