% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_global_totalResourcesValueBox.R
\name{totalResources_serv}
\alias{totalResources_serv}
\title{Global Module: totalResourcesValueBox (Server Code)}
\usage{
totalResources_serv(id, sel_categories)
}
\arguments{
\item{id}{The namespace-id for this Module.}

\item{sel_categories}{Vector of selected categories.}
}
\value{
Server Code for the Module
}
\description{
Global Module: totalResourcesValueBox (Server Code)
}
