% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_global_totalResourcesValueBox.R
\name{totalResourcesValueBox}
\alias{totalResourcesValueBox}
\title{Title}
\usage{
totalResourcesValueBox(id, width = 2)
}
\arguments{
\item{id}{Session id}
}
\value{
valueBox with number of resources that will be downloaded
}
\description{
Title
}
