% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_connect_utility.R
\name{try_project_connection}
\alias{try_project_connection}
\title{Try to get some data from the selected project}
\usage{
try_project_connection(connection, project)
}
\arguments{
\item{connection}{an object as returned by \code{idaifieldR::connect_idaifield()}}

\item{project}{the project name}
}
\value{
TRUE if succesfull, message (character) if not
}
\description{
Try to get some data from the selected project
}
