% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_utility.R
\name{ui_options_title}
\alias{ui_options_title}
\title{tagList for the Title of Selection- and Plot-Options}
\usage{
ui_options_title(type = "plot")
}
\arguments{
\item{type}{chr. "plot" or "selection"}
}
\value{
\code{tagList()} with Icon and Title
}
\description{
tagList for the Title of Selection- and Plot-Options
}
