% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_global_categorySelection.R
\name{uiCategorySelector}
\alias{uiCategorySelector}
\title{Global Module: categorySelector (UI)}
\usage{
uiCategorySelector(id)
}
\arguments{
\item{id}{The namespace-id for this Module.}
}
\value{
\code{htmlOutput()}-Element of the category selector.
}
\description{
Global Module: categorySelector (UI)
}
