% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_global_helpers.R
\name{uid_by_operation}
\alias{uid_by_operation}
\title{Returns the UIDs of entries belonging to the variable in question}
\usage{
uid_by_operation(filter_operation = "all", index = react_index())
}
\arguments{
\item{index}{}
}
\description{
Returns the UIDs of entries belonging to the variable in question
}
