% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_global_layerSelection.R
\name{uiLayerSelector}
\alias{uiLayerSelector}
\title{Global Module: layerSelector (UI)}
\usage{
uiLayerSelector(id)
}
\arguments{
\item{id}{The namespace-id for this Module.}
}
\value{
\code{htmlOutput()} of the layer selector.
}
\description{
Global Module: layerSelector (UI)
}
