% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_global_periodSelection.R
\name{uiPeriodSelector}
\alias{uiPeriodSelector}
\title{Global Module: periodSelector (UI)}
\usage{
uiPeriodSelector(id)
}
\arguments{
\item{id}{The namespace-id for this Module.}
}
\value{
\code{htmlOutput()} of period selector
}
\description{
Global Module: periodSelector (UI)
}
