% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_dating.R
\name{add_all_periods_to_all_contexts}
\alias{add_all_periods_to_all_contexts}
\title{add_all_periods_to_all_contexts}
\usage{
add_all_periods_to_all_contexts(data)
}
\arguments{
\item{data}{data.frame with the resources to be transformed}
}
\value{
data.frame where each context has an empty row for earch period (to make it possible to use the same legend in plotly)
}
\description{
add_all_periods_to_all_contexts
}
