% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_utility.R
\name{addLRPopover}
\alias{addLRPopover}
\title{Title}
\usage{
addLRPopover(input)
}
\arguments{
\item{input}{the ui object that the popover should be attached to}
}
\value{
same input with a tooltip/popover to inform of Load Resources button behaviour
}
\description{
Title
}
