% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_utility.R
\name{addLRPopover}
\alias{addLRPopover}
\title{Add the \emph{Load Resources}-Popover to UI-Element}
\usage{
addLRPopover(input)
}
\arguments{
\item{input}{The UI-object that the popover should be attached to}
}
\value{
Same input with a tooltip/popover to inform of \emph{Load Resources}-button behaviour
}
\description{
Add the \emph{Load Resources}-Popover to UI-Element
}
