% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_global_get_data.R
\name{alt_gather_trenches}
\alias{alt_gather_trenches}
\title{Alternative to gather trenches for milQuant}
\usage{
alt_gather_trenches(uidlist)
}
\description{
Alternative to gather trenches for milQuant
}
