% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_activity.R
\name{db_activity_tab}
\alias{db_activity_tab}
\title{Tab-Module for Activity (UI)}
\usage{
db_activity_tab(id, tabname)
}
\arguments{
\item{id}{The namespace-id for this Module.}

\item{tabname}{Name of the tab used to link this module.}
}
\value{
UI for the Module
}
\description{
Tab-Module for Activity (UI)
}
