% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_home_db_overview.R
\name{db_overview_tab}
\alias{db_overview_tab}
\title{TabPanel-Module for Project Overview (UI)}
\usage{
db_overview_tab(id, tabname)
}
\arguments{
\item{id}{The namespace-id for this Module.}

\item{tabname}{Name of the tab used to link this module.}
}
\value{
UI for the Module
}
\description{
TabPanel-Module for Project Overview (UI)
}
