% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_global_downloadModule.R
\name{downloadPlotButtons}
\alias{downloadPlotButtons}
\title{Title}
\usage{
downloadPlotButtons(id)
}
\arguments{
\item{id}{Session id}
}
\value{
UI download plot button
}
\description{
Title
}
