% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_global_categorySelection.R
\name{generateCategorySelector}
\alias{generateCategorySelector}
\title{Global Module: categorySelector (Server Code)}
\usage{
generateCategorySelector(id, inputid, parent = "Find", selected = "auto")
}
\arguments{
\item{id}{The namespace-id for this Module.}

\item{inputid}{The \code{inputId} that should be used for the UI-Element.
Specified here to avoid \code{ns()}-problems and keep track of the value easier.}

\item{parent}{\emph{parent}-category of the resources that should be selectable in the input.}

\item{selected}{"auto", "all", "none" or a vector of categories the should be preselected}
}
\value{
Server Code for the Module
}
\description{
Global Module: categorySelector (Server Code)
}
