% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_global_categorySelection.R
\name{generateCategorySelector}
\alias{generateCategorySelector}
\title{Title}
\usage{
generateCategorySelector(id, inputid, parent = "Find", selected = "auto")
}
\arguments{
\item{inputid}{}
}
\description{
Title
}
