% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_global_periodSelection.R
\name{generatePeriodSelector}
\alias{generatePeriodSelector}
\title{Title}
\usage{
generatePeriodSelector(id, inputid)
}
\arguments{
\item{id}{Session id}

\item{inputid}{the inputid the selector should have}
}
\value{
server code for period selector
}
\description{
Title
}
