% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_global_get_data.R
\name{get_resources}
\alias{get_resources}
\title{get_resources}
\usage{
get_resources(resource_category = "Pottery", fields = "all")
}
\arguments{
\item{resource_category}{}
}
\description{
get_resources
}
