% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_get_data.R
\name{get_resources}
\alias{get_resources}
\title{Title}
\usage{
get_resources(
  resource_category = "Pottery",
  fields = "all",
  liesWithinLayer = NULL,
  prep_for_shiny = TRUE
)
}
\arguments{
\item{resource_category}{see milQuant_cats}

\item{fields}{fields to get from the db}

\item{liesWithinLayer}{if a vector of layers, will be filtered by finds from those layers according to index}

\item{prep_for_shiny}{should prep_for_shiny() be run?}
}
\value{
depending on prep_for_shiny, either a list or a data.frame
}
\description{
Title
}
