% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_tab_home.R
\name{make_home_tab}
\alias{make_home_tab}
\title{Home-Tab / Landing Page tabitem (UI, not a module)}
\usage{
make_home_tab(tabname = "tab_home")
}
\arguments{
\item{tabname}{Name of the tab used to link this module.}
}
\value{
UI for the Module
}
\description{
Home-Tab / Landing Page tabitem (UI, not a module)
}
