% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_global_downloadModule.R
\name{makeDownloadPlotHandler}
\alias{makeDownloadPlotHandler}
\title{Title}
\usage{
makeDownloadPlotHandler(id, dlPlot)
}
\arguments{
\item{id}{Session id}

\item{dlPlot}{the plot to be downloaded}
}
\value{
server code for download button
}
\description{
Title
}
