% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_finds_aoristic.R
\name{mod_aoristic_finds_ui}
\alias{mod_aoristic_finds_ui}
\title{Title}
\usage{
mod_aoristic_finds_ui(id, tabname)
}
\arguments{
\item{id}{Session id}

\item{tabname}{Name of the tab used to link it}
}
\value{
The Tab with Aoristic evaluation of Chronology for find groups
}
\description{
Title
}
