% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_finds_barchart.R
\name{mod_barchart_finds_ui}
\alias{mod_barchart_finds_ui}
\title{Tab-Module for \emph{Find}-Bar Charts (UI)}
\usage{
mod_barchart_finds_ui(id, tabname)
}
\arguments{
\item{id}{The namespace-id for this Module.}

\item{tabname}{Name of the tab used to link this module.}
}
\value{
UI for the Module
}
\description{
Tab-Module for \emph{Find}-Bar Charts (UI)
}
