% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_finds_quant.R
\name{mod_finds_quant_ui}
\alias{mod_finds_quant_ui}
\title{Title}
\usage{
mod_finds_quant_ui(id, tabname)
}
\arguments{
\item{id}{Session id}

\item{tabname}{Name of the tab used to link it}
}
\value{
A tab which displays the Total value from all Quantifications
}
\description{
#TODO: add up all individual counts / weights in case total does not exist;
but could also do that more reasonable by exporting and overwriting in the db itself.
}
