% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_loomweights_hist.R
\name{mod_loomweights_hist_ui}
\alias{mod_loomweights_hist_ui}
\title{Tab-Module for Loomweight-Historgram (UI)}
\usage{
mod_loomweights_hist_ui(id, tabname)
}
\arguments{
\item{id}{The namespace-id for this Module.}

\item{tabname}{Name of the tab used to link this module.}
}
\value{
UI for the Module
}
\description{
Tab-Module for Loomweight-Historgram (UI)
}
