% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_pottery_QA.R
\name{mod_pottery_QA_ui}
\alias{mod_pottery_QA_ui}
\title{Title}
\usage{
mod_pottery_QA_ui(id, tabname)
}
\arguments{
\item{id}{Session id}

\item{tabname}{Name of the tab used to link it}
}
\value{
A tab which displays the Pottery Quantifications
}
\description{
If I make it only slightly customizable, this could actually be a better
generalized version for all quant tabs
}
