% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_pottery_QB.R
\name{mod_pottery_QB_ui}
\alias{mod_pottery_QB_ui}
\title{Title}
\usage{
mod_pottery_QB_ui(id, tabname)
}
\arguments{
\item{id}{Session id}

\item{tabname}{Name of the tab used to link it}
}
\value{
Tab with evaluation of the pottery quant B forms by period
}
\description{
Title
}
\section{TODO might want to make it possible to do this with absolute}{
values replaced from the periods
}

