% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_pottery_single.R
\name{mod_pottery_single_ui}
\alias{mod_pottery_single_ui}
\title{Title}
\usage{
mod_pottery_single_ui(id, tabname)
}
\arguments{
\item{id}{Session id}

\item{tabname}{Name of the tab used to link it}
}
\value{
Tab with charts specifically for pottery
}
\description{
#TODO make this less dynanic and more curated
}
