% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_quants.R
\name{mod_quants_serv}
\alias{mod_quants_serv}
\title{Tab-Module for Single Quantification Forms (Server Code)}
\usage{
mod_quants_serv(id, resource_category = "Brick_Quantification")
}
\arguments{
\item{id}{The namespace-id for this Module.}

\item{resource_category}{The category of \emph{resource}s to be used in this module.}
}
\value{
Server Code for the Module
}
\description{
Tab-Module for Single Quantification Forms (Server Code)
}
