% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_quants.R
\name{mod_quants_ui}
\alias{mod_quants_ui}
\title{Tab-Module for Single Quantification Forms (UI)}
\usage{
mod_quants_ui(id, tabname)
}
\arguments{
\item{id}{The namespace-id for this Module.}

\item{tabname}{Name of the tab used to link this module.}
}
\value{
UI for the Module
}
\description{
Tab-Module for Single Quantification Forms (UI)
}
