% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_search.R
\name{mod_search_ui}
\alias{mod_search_ui}
\title{Title}
\usage{
mod_search_ui(id, tabname = NULL)
}
\arguments{
\item{id}{Session id}

\item{tabname}{Name of the tab used to link it}
}
\value{
Tab with a table that enables full text search of text inputs
}
\description{
Title
}
