% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_utility.R
\name{mq_spinner}
\alias{mq_spinner}
\title{mq_spinner}
\usage{
mq_spinner(object)
}
\arguments{
\item{object}{ui objects that needs a spinner}
}
\value{
ui object witht spinner
}
\description{
mq_spinner
}
