% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_dating.R
\name{multiply_resources_by_period_range}
\alias{multiply_resources_by_period_range}
\title{multiply_resources_by_period_range}
\usage{
multiply_resources_by_period_range(data, split = TRUE)
}
\arguments{
\item{data}{data.frame with the resources to be transformed}

\item{split}{TRUE/FALSE. A way out to not multiply while still calling the function}
}
\value{
data.frame with duplicated rows and counts devided by nr of duplicates
}
\description{
#TODO use this from chrongler
}
