% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_dating.R
\name{period_filter}
\alias{period_filter}
\title{apply the period filter if the config is milet}
\usage{
period_filter(find_df, is_milet = FALSE, selector = NULL)
}
\arguments{
\item{find_df}{df of find resources to select for periods}

\item{is_milet}{on project milet?}

\item{selector}{the list from the associated input}
}
\value{
the df, selected
}
\description{
apply the period filter if the config is milet
}
