% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_global_displayPlotDataTable.R
\name{plotDataTable_server}
\alias{plotDataTable_server}
\title{Global Module: clickDataTable (Server Code)}
\usage{
plotDataTable_server(id, resources, click_data, x, customdata)
}
\arguments{
\item{id}{The namespace-id for this Module.}

\item{resources}{Reactive data of all \emph{resources} to be displayed in the table.}

\item{click_data}{The reactive object returned by plotly click data.}

\item{x}{First variable to select for.}

\item{customdata}{Second variable to select for.}
}
\value{
Server Code for the Module
}
\description{
Global Module: clickDataTable (Server Code)
}
