% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_global_displayPlotDataTable.R
\name{plotDataTable_server}
\alias{plotDataTable_server}
\title{Title}
\usage{
plotDataTable_server(id, resources, click_data, x, customdata)
}
\arguments{
\item{customdata}{}
}
\description{
Title
}
