% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_global_displayPlotDataTable.R
\name{plotDataTable_ui}
\alias{plotDataTable_ui}
\title{Global Module: clickDataTable (UI)}
\usage{
plotDataTable_ui(id)
}
\arguments{
\item{id}{The namespace-id for this Module.}
}
\value{
A \code{fluidRow()} with: h2, column selector input, and datatable output.
}
\description{
Global Module: clickDataTable (UI)
}
