% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_global_helpers.R
\name{prep_for_shiny}
\alias{prep_for_shiny}
\title{Prep for shiny}
\usage{
prep_for_shiny(data, reorder_periods = reorder_periods)
}
\arguments{
\item{reorder_periods}{}
}
\description{
Prep for shiny
}
