% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_utility.R
\name{remove_na_cols}
\alias{remove_na_cols}
\title{Helper to remove columns that are empty}
\usage{
remove_na_cols(data)
}
\arguments{
\item{data}{data.frame from which columns with all na-vals should be removed}
}
\value{
data.frame
}
\description{
Helper to remove columns that are empty
}
