% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_global_totalResourcesValueBox.R
\name{totalResourcesValueBox}
\alias{totalResourcesValueBox}
\title{Global Module: totalResourcesValueBox (UI)}
\usage{
totalResourcesValueBox(id, width = 2)
}
\arguments{
\item{id}{The namespace-id for this Module.}

\item{width}{Width of the \code{valueBox()}}
}
\value{
\code{valueBox()} with number of resources selected as gathered from Index.
}
\description{
Global Module: totalResourcesValueBox (UI)
}
