% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_ui_elements.R
\name{ui_options_title}
\alias{ui_options_title}
\title{tagList for Title of Plot/Selection options}
\usage{
ui_options_title(type = "plot")
}
\arguments{
\item{type}{plot or selection}
}
\value{
tagList with icon and title
}
\description{
tagList for Title of Plot/Selection options
}
